newstreamer-main/requirements.txt
# NewStreamer 依赖包
# Python >= 3.8

# 数据处理
pandas>=1.3.0
numpy>=1.21.0

# JSON处理
ujson>=5.1.0

# 网络请求
requests>=2.26.0
websocket-client>=1.2.0

# Redis缓存
redis>=4.0.0

# 日志
loguru>=0.6.0

# 测试
pytest>=7.0.0
pytest-cov>=3.0.0

# 类型检查
typing-extensions>=4.0.0

# 开发工具
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

